Moving game steps to modules
