+++ lib/earth_server/utils/communication.ex
@@ -1,0 +1,6 @@
+defmodule EarthServer.Utils.Communication do
+  @spec communication() :: EarthServer.Communication.t()
+  def communication do
+    Application.get_env(:earth_server, :communication)
+  end
+end
